Changing abstract to interface and adding namespace.
