+++ Brains/Assets/Scripts/Abstracts & Interfaces/Character.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5e45a282137a7ea4ab474d2ed5cf369a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
