--- Brains/Assets/Scripts/Abstracts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 81454693b2ce0fb4a86d387eef4f5eca
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
