+++ Brains/Assets/Scripts/Player/Player.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 01a5bd295e45e67409c6c544c853d51f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
